removing files, renaming files, moving tests
